server/api/requirements.txt
aniso8601==7.0.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Cors==3.0.10
Flask-GraphQL==2.0.1
graphene==2.1.9
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.5
Pillow==9.3.0
promise==2.3
requests==2.28.1
Rx==1.6.1
six==1.16.0
torch==1.13.0
torchvision==0.14.0
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.11.0
